// src/app/models/employee.ts
export class Employee{
    url: string;
    id: number;
    username: string;
    first_name: string;
    last_name: string;
    email: string;
    is_active: boolean;
    date_joined: string;
    avatar :string;
    check_params:boolean;
    check_contacts: boolean;
    param_list: string;
    contacts: string;
    description: string;
    rating: number;
}